MailChimp plugin minor changes and refactoring

+++ src/Plugins/Nop.Plugin.Misc.MailChimp/Services/SubscriptionEventQueueingService.cs
@@ -1,54 +1,53 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using Nop.Core.Data;
 using Nop.Plugin.Misc.MailChimp.Data;
 
-namespace Nop.Plugin.Misc.MailChimp.Services {
-    public class SubscriptionEventQueueingService : ISubscriptionEventQueueingService {
+namespace Nop.Plugin.Misc.MailChimp.Services
+{
+    public class SubscriptionEventQueueingService : ISubscriptionEventQueueingService
+    {
         private const string QUEUE_ALL_RECORDS = "INSERT INTO MailChimpEventQueueRecord (Email, IsSubscribe) SELECT Email, 1 FROM NewsLetterSubscription WHERE Active = 1";
 
         private readonly IRepository<MailChimpEventQueueRecord> _repository;
         private readonly MailChimpObjectContext _context;
 
-        public SubscriptionEventQueueingService(IRepository<MailChimpEventQueueRecord> repository, MailChimpObjectContext context) {
+        public SubscriptionEventQueueingService(IRepository<MailChimpEventQueueRecord> repository, MailChimpObjectContext context)
+        {
             _repository = repository;
             _context = context;
         }
-
-        #region Implementation of ISubscriptionEventQueueingService
 
         /// <summary>
         /// Deletes the specified record.
         /// </summary>
         /// <param name="record">The record.</param>
-        public void Delete(MailChimpEventQueueRecord record) {
+        public virtual void Delete(MailChimpEventQueueRecord record)
+        {
+            if (record == null)
+                throw new ArgumentNullException("record");
+
             _repository.Delete(record);
         }
 
         /// <summary>
-        /// Enqueues the specified record.
+        /// Inserts the specified record.
         /// </summary>
         /// <param name="record">The record.</param>
-        public void Enqueue(MailChimpEventQueueRecord record) {
+        public virtual void Insert(MailChimpEventQueueRecord record)
+        {
+            if (record == null)
+                throw new ArgumentNullException("record");
+
             _repository.Insert(record);
         }
-
-        /// <summary>
-        /// Reads the next.
-        /// </summary>
-        /// <returns></returns>
-        public MailChimpEventQueueRecord ReadNext() {
-            var query = from r in _repository.Table
-                        orderby r.Id
-                        select r;
-
-            return query.FirstOrDefault();
-        }
-
+        
         /// <summary>
         /// Queues all subscriptions.
         /// </summary>
-        public void QueueAll() {
+        public virtual void QueueAll()
+        {
             //NOTE: While I dislike executing straight SQL from C#, I think straight SQL is the best solution for this particular action.
             // My logic is that #1 this query takes no arguments, it is a batch query and #2 this is potentially a very big operation.
             // Loading all of the NewsLetterSubscriptions, iterating them, converting them, and submitting the changes just seems like we would not be taking advantage of our tools (SQLServer).
@@ -61,7 +60,8 @@ namespace Nop.Plugin.Misc.MailChimp.Services {
         /// </summary>
         /// <param name="subscribeOnly">if set to <c>true</c> [subscribe only].</param>
         /// <returns></returns>
-        public IList<MailChimpEventQueueRecord> ReadList(bool subscribeOnly) {
+        public virtual IList<MailChimpEventQueueRecord> ReadList(bool subscribeOnly)
+        {
             var query = from r in _repository.Table
                         orderby r.Id
                         where r.IsSubscribe == subscribeOnly
@@ -69,7 +69,5 @@ namespace Nop.Plugin.Misc.MailChimp.Services {
 
             return query.ToList();
         }
-
-        #endregion
     }
 }
